chore: upgrade PHPStan to level 7
Fixed return type inconsistency:
- AutoscaleManager::getCooldownRemaining(): Cast max() result to int
  to ensure int return type instead of float|int

Level 7 error resolved. Baseline remains at 3 unavoidable errors.

Tests: 84 passed (169 assertions)

+++ src/Manager/AutoscaleManager.php
@@ -388,6 +388,6 @@ final class AutoscaleManager
         $lastScale = $this->lastScaleTime[$key];
         $elapsed = $lastScale->diffInSeconds(now());
 
-        return max(0, $cooldownSeconds - $elapsed);
+        return (int) max(0, $cooldownSeconds - $elapsed);
     }
 }
